requirements.txt
tkinter
requests
Pillow
openai
bing_image_urls
Stability_sdk